Put testTakeMarketingScreenshots in L10nSuiteSnapshotTests
